+++ template/legostore/src/main/java/rc/voxxed/legostore/model/DeliveryInfo.java
@@ -1,22 +1,21 @@
 package rc.voxxed.legostore.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class DeliveryInfo {
     private LocalDate deliveryDate;
-    private BigDecimal deliveryFee;
+    private int deliveryFee;
     private boolean inStock;
 
     public DeliveryInfo(LocalDate deliveryDate,
-                        BigDecimal deliveryFee,
+                        int deliveryFee,
                         boolean inStock) {
         this.deliveryDate = deliveryDate;
         this.deliveryFee = deliveryFee;
         this.inStock = inStock;
     }
 
-    public BigDecimal getDeliveryFee() {
+    public int getDeliveryFee() {
         return this.deliveryFee;
     }
